Add HVARFNER preset and make BOTORCH reference it
The BOTORCH preset delegates to BoTorch utilities that currently implement
the approach from Hvarfner et al. (2024). Since BoTorch's defaults may
change over time, the fixed Hvarfner method now lives in its own preset
module, and BOTORCH simply aliases it.
