+++ docs/references.bib
@@ -17,6 +17,13 @@
  doi = {10.26434/chemrxiv.10001986/v2}
 }
 
+@inproceedings{Hvarfner2024,
+ author = {Hvarfner, Carl and Hellsten, Erik Orm and Nardi, Luigi},
+ title = {Vanilla Bayesian Optimization Performs Great in High Dimensions},
+ booktitle = {International Conference on Machine Learning},
+ year = {2024}
+}
+
 @article{Shields2021,
  author = {Shields, Benjamin J. and Stevens, Jason and Li, Jun and Parasram, Marvin and Damani, Farhan and Alvarado, Jesus I. Martinez and Janey, Jacob M. and Adams, Ryan P. and Doyle, Abigail G.},
  title = {Bayesian reaction optimization as a tool for chemical synthesis},
