+++ tests/test_gp.py
@@ -196,8 +196,9 @@ def test_invalid_components():
 
 
 @pytest.mark.parametrize("multitask", [False, True], ids=["single-task", "multi-task"])
-def test_botorch_preset(multitask: bool):
-    """The BoTorch preset exactly mimics BoTorch's behavior."""
+@pytest.mark.parametrize("preset", ["BOTORCH", "HVARFNER"], ids=["botorch", "hvarfner"])
+def test_botorch_preset(multitask: bool, preset: str):
+    """The BoTorch/Hvarfner presets exactly mimic BoTorch's behavior."""
     if multitask:
         sp = searchspace_mt
         data = measurements_mt
@@ -206,7 +207,7 @@ def test_botorch_preset(multitask: bool):
         data = measurements
 
     active_settings.random_seed = 1337
-    gp = GaussianProcessSurrogate.from_preset("BOTORCH")
+    gp = GaussianProcessSurrogate.from_preset(preset)
     gp.fit(sp, objective, data)
     posterior1 = gp.posterior_stats(data)
